feat(json-init): add a --with-json-config=path/to/config.json flag for passing json to init
This exists for hosted services that want to edit the toml mechanically.
The format is basically the json equivalent of `[workspace.metadata.dist]`,
wrapped in a `workspace` field:

```json
{
    "workspace": {
        "cargo-dist-version": "0.1.0-prerelease.2",
        "ci": ["github"],
        "installers": ["shell", "powershell"],
        "targets": ["x86_64-unknown-linux-gnu", "x86_64-apple-darwin", "x86_64-pc-windows-msvc", "aarch64-apple-darwin"]
    }
}
```

The given values will complete overwrite the `[workspace.metadata.dist]` section,
so if any fields are omitted they will be deleted from the Cargo.toml.

All the interactive logic that `init` supports will be ignored in this mode
(this is *NOT* equivalent to --yes, which runs the logic and just accepts every prompt).

generate-ci will be run afterwards as usual (disable with --no-generate-ci as usual).

fixes #278