chore: add misty's project to the gallery
This also reworks the gallery code to properly have a global lock between all Repos
since we cd into the repo before running a test, which is global mutable state.
I did this by requiring you to define a global lock for Tools that must be passed
by-ref into each TestContextLock::new call, and using that as THE global lock
(which is fine, Tools getting initialized once for each Repo was jank anyway).