feat(features): add the ability to specify features
This adds the following configs to metadata.dist:

```toml
features = ["blah", "blargh"]
no-default-features = true
all-features = true
```

If applied to [workspace.metadata.dist] it will be applied to all packages

If applied to [package.metadata.dist] it will be applied to just that package (overriding the workspace value if it exists).

If not all packages end up with the same feature flags then precise-builds=true will be forced on.