fix: properly handle multiple subscribers to a binary
The previous code accidentally assumed there was only one subscriber to each binary,
and so if there were multiple ones one of them would arbitrarily win and get to copy
the binary, while the other wouldn't silently get dropped.

The exact way it got dropped also bypassed the safety checks against this, alas.

The fix is simple: make the subscribers a list, instead of a single entry.