fix(sysdeps): handle empty Linux packages
We've seen a minor UI issue with Linux dependencies. I anticipated catching
the error status output from `dpkg -S` and reacting to that, but we were
still getting cases where not finding the package and accidentally
returning an empty string as the "package". This checks for that case and
properly handles it.