feat(installer): add additional shell config
~/.profile isn't sourced in many circumstances; we need to write this
to a large number of locations to make sure it gets considered by bash and
zsh under the largest number of circumstances.

We have three sets of configuration files, covering generic POSIX, bash, and zsh.
We pick the first file that exists in each set, and if none exists, we create the
first in the list.

Fixes #547.