feat(dockerimage): add a docker image installer
This creates a Local (as in one-per-platform) docker installer (which only runs for x64 gnu linux)
that builds a docker image that vendors the binaries, exporting to a tarball.

This does not currently setup docker buildx in CI (so might fail).
This does not currently support publishing the docker image anywhere.