fix(extra artifacts): avoid inappropriate builds
We had an issue where we would still build extra artifacts even if they'd been
filtered out of the target artifact list. This fixes that.

The issue was caused by the separate logic we had to use to add these builds to
the graph. Because each of these builds produces multiple artifacts, we weren't
able to add them in the same place that other global artifacts are added. However,
that meant we weren't taking into account the list of target artifacts - we were
unconditionally building them, even if we didn't want them.

Also fixes an issue where they were assigned to the local build array instead
of the global build array, though the impact of that bug was minor.

Fixes #658.