fix(source): skip source tarballs if no commits
When deciding whether to generate a source tarball, we check whether the user
is in a git repo but we don't check if that git repo has any commits. This is
an edge case, but very possible: a user might have run `git init`, added some
code, but not yet committed anything.

Fixes #695.