feat: test installers
This tests our own generated installers instead of the installers that were
build for the version we're fetching. It asks axoupdater what version it plans
to update to, then generates new installers and updates the version we're fetching
to that instead of whatever cargo-dist would have generated. Using a new testing
feature from axoupdater, we then substitute these new installers for the ones it
would have downloaded from the new release.