feat: clean up force install behaviour
Instead of always overriding after setting the path, this places it
directly inside each of the branches. This lets us customize the values to
be more appropriate to each branch, fixing an issue where `/bin` would be added to
paths that shouldn't have it.

In addition, it adds a new check to several branches for if `$HOME` is contained
within and substitutes the literal value for the variable in the `expr` values, and
avoids overwriting the configured values if the forced values are already equivalent
to the defaults.