feat: migrate from flat to hierarchical installs (.sh)
Before this commit, we had two different installation layouts:

* Cargo home produced an FHS-style installation layout, with binaries
  installed into the `bin` directory within an install prefix.
* Env subdir and home subdir produced a flat installation layout, with
  binaries and all other files placed in a single directory.

Going forward, it will be advantageous for us to have a unified strategy where we
can predict the layout of an install directory no matter which path we're following.

This doesn't just change the default, but also performs a migration: any binaries that
this installer would install that exist in the old default path are migrated to the new
path before the installation itself begins.