factor platform code into shared unit
This is a significantly reduced scope implementation of #793 that takes all the rich info
and shoves it back into the shape of ExecutableZipFragment so all the installers can remain the same.
However in doing so it introduces much better support for aliases and fallbacks, and centralizes the logic.

As a result several missing aliases are added, and fallback is much more maintainable.