feat: allow configuring global runner
We already parameterized the "global task" runner, but hardcoded it to always
use a specific GitHub Linux runner. This updates it to now use the "global" key
in the pre-existing `github-custom-runners` table, if it's set. It also updates
`plan`, the one non-parameterized task, to use this runner too.

Updates the custom runner test to exercise this new feature.