feat: autodetect buildjet runners to use their backend for rust cache
This is an alternative based on #1119 which instead of introducing a coarse-grained
config, adds auto-detection for each individual runner based on the name containing
"buildjet". This then feeds down through the github matrix, so the cache is hosted
precisely per-platform.

Big thanks to @arlyon for the initial implementation and insights!