feat: cache cargo-dist binary in global builds
Currently, we fetch or build the cargo-dist binary every single time. If we're
building for git, this gets repetitive and slow. This speeds it up by uploading
the binary we fetched to the build artifacts, where we can fetch it back for
later runs without rebuilding from source.

This doesn't affect local builds, only things which use the global build runner.