feat: support C dynamic libraries
We had cdylibs available from axoproject/`cargo metadata` in the past,
but we ignored them. This pipes them into the buildsystem and treats
them as binaries like any others. For most of the layers of cargo-dist,
we have a single set of "binaries" but we now type them so that we can
keep track of which are executables and which are libraries. We separate
them back out at the time that we create installers to simplify the
process of deciding where they get installed to.

By default, these aren't packaged or installed in order to avoid catching
people by surprise. However, they can be enabled easily via the new
`package-cdylibs` and `install-cdylibs` config flags.

Homebrew has full first-class support for libraries, which we use here.
Shell and PowerShell currently install them to the same path as binaries;
this will change with new default prefixes in the future. npm projects
don't change since they never did any form of organized installation, and
just dumped the entire contents of the release tarball into a private path.