feat: track glibc via runtime requirements
So far, we have one runtime requirement expressed directly in our
installers: the glibc version used to build. At the moment, this is
hardcoded based on the value in the default runner, but we'd like to
have it be determined at buildtime and then used in the installer as
appropriate.

This adds that tracking. At buildtime, we now add to the `SystemInfo`
field to track system environment information specific to the build
environment. We should expand this in time, but at the moment we track
builder glibc version and macOS version. We then collate that
information in the installer to calculate the safest minimum glibc
version and inject that into the installers rather than hardcoding.