fix: centralize repository url checks and give more detailed diagnostic
This fixes a longstanding hole in init where it tried to eagerly validate the repository-url setting but was too early in the process to actually have the info to do that validation. While adding details to the errors I realized that the eager validation is actually completely counter-productive and it can just be deleted!

As soon as we run generate we will also do this validation but with full context, and, importantly, by not doing the eager validation we don't delete all the config work the user just did. The config gets saved, and then we error out, so you're nice and 'checkpointed'.