chore: rework RuntimeConditions
This turns RuntimeConditions into a flat struct with Options, reflecting the fact that a list of enums is really hard to work with and can express confusing ideas.

This also makes PlatformSupport::new compute RuntimeConditions from the current system and linkage info, working bottom-up from binary conditions up to archive and platform conditions.

It also adds a new bridging API conflated_runtime_conditions to be paired with ragments users.

The implementation is currently unused and missing a few key parts at the bottom for glibc version, which largely exist in Misty's PR.