feat: macOS codesigning
This adds support for codesigning on macOS, similar to the ssl.com
feature on Windows. In the current revision, we only perform
signing with a certificate - notarization will come in a subsequent
feature.

Similar to ssl.com, the only feature in our config is the feature
controlling whether or not to sign - the rest is handled via the
environment. The user needs to provide three fields:

* A pkcs12 certificate, base64-encoded.
* The passphrase for that certificate.
* The name of the codesigning identity belonging to that certificate.

If all three are present, we'll codesign any binaries and libraries
during the local build process.

This uses the builtin macOS tooling, which means it can only be performed
on a macOS host. (That's not too different from how we can only do the
Windows codesigning on a Windows host.) Because interacting with the
keychain is slightly wonky on something like a hosted runner, we've got
a slightly complex process that takes a few steps. (This process is
inspired by official documentation from GitHub:
https://docs.github.com/en/actions/use-cases-and-examples/deploying/installing-an-apple-certificate-on-macos-runners-for-xcode-development)

* We create a fresh keychain database in a temporary directory, with
  a randomized password under our control.
* We unlock that keychain and configure it to remain unlocked for the
  duration of the signing process.
* We set that keychain as the default.
* We import our certificate into that keychain, and configure it so
  that it can only be used by `codesign`, `security` and `productsign`.
* We configure the key to be used for signing.
* We actually sign the binary.