feat: support app-specific force install
This supports a new app-specific force install environment variable.
We use the app's name, uppercased and with `-` replaced by `_`.
For example, if the app name is `akaikatana-repack`, the new force install
environment variable becomes `AKAIKATANA_REPACK_FORCE_INSTALL_DIR`.

This supports the new variable alongside the old one. In the longterm, we
could consider removing the old one after a transition period where both are
supported.

An axolotlsay PR to use the new environment variable is here:
https://github.com/axodotdev/axoupdater/pull/169

refs #1375.