feat: unmanaged installs
The new "unmanaged" installation mode allows for something close
to "untar and move on" installations. It disables modifying the
PATH and creation of environment files, and ensures all files are
installed into a single flat directory no matter what other
configuration is in place. It also disables installing a standalone
updater and the creation of an install receipt.

The new "don't install an updater/receipt" behaviour is also
controllable behind a new public environment variable,
$APPNAME_DISABLE_UPDATE.

Fixes #1400.