fix(ci): always serialize upload_files
While it's usually an error if we received a totally empty set of
artifacts, it's possible in the global build job under a very
specific set of conditions:

* There are no installers, *and*
* The source tarball was disabled

In this case we need to make sure we still provide an upload_files
array in the manifest so that we can always index into it.

Fixes #1387.