fix: standalone updater installation
This was broken in a refactor in 19af85324be25d023b2bac291aa105f081d7da30.
That commit updated how we fetch platform-specific information, but it missed
that the `updater` field in the new type is differently-shaped than the field
from the new type. Specifically, the new `archive` we're working on has an
`updater` field that's an *index into a table of updaters* intead of an
*updater struct* - so trying to read its `id` and `bin` fields was doomed
to failure.

Fixes #1443.