feat: allow installer domain to be overridden
Previously, we allowed overriding the full URL to the installer path -
that is, everything prior to the actual installer filename. This isn't
flexible enough for people who are mirroring or proxying github.com, and
who want to be able to set a single environment variable that can work
for any version of a program.

This adds two environment variables, `INSTALLER_GITHUB_BASE_URL` and
`INSTALLER_GITHUB_ENTERPRISE_BASE_URL`, which allow overriding just the
domain. They do the same thing for these installers, but will have
slightly different behaviour for axoupdater.

Fixes #1495.