feat: legalize [package] and [workspace]
Previously, we disallowed both [package] and [workspace] in a
single dist-workspace.toml. However, we want to allow this for
generic/non-Rust projects so that these users only need a single
metadata file, not two. This required a few changes to axoproject
in order to make sure we structure the data we're returning
properly.