Vendor axo dependencies into workspace
Axo is unfortunately defunct and all of their projects `cargo-dist`
depends on are also unmaintained. We don't intend to perform any changes
to `dist` other than keeping its dependencies up-to-date, but this can
will be simplified by bringing all of the unmaintained crates into a
single repository.

Vendor all Axo dependencies at the version `dist` depended on. We remove
a few unused files, `.github` and `Cargo.lock`, but otherwise bring
these in unmodified. The exception is `axoprocess`, which uses a simple
workspace with two sub-packages. We trim this down to just the package
used by `dist`.